requirements.txt
aiogram>=3.4.1,<4.0.0
Pillow>=10.2.0
